+++ modules/Attendance/TeacherCompletion.php
@@ -2,50 +2,46 @@
 
 require_once 'ProgramFunctions/TipMessage.fnc.php';
 
-// set date
-if ( isset( $_REQUEST['month_date'] )
-	&& isset( $_REQUEST['day_date'] )
-	&& isset( $_REQUEST['year_date'] ) )
-{
-	$date = RequestedDate(
-		$_REQUEST['year_date'],
-		$_REQUEST['month_date'],
-		$_REQUEST['day_date']
-	);
-}
-else
-{
-	$_REQUEST['day_date'] = date('d');
-	$_REQUEST['month_date'] = date('m');
-	$_REQUEST['year_date'] = date('Y');
+// Set date.
+$date = RequestedDate( 'date', DBDate(), 'set' );
 
-	$date = $_REQUEST['year_date'] . '-' . $_REQUEST['month_date'] . '-' . $_REQUEST['day_date'];
-}
+DrawHeader( ProgramTitle() );
+$categories_RET = DBGet( "SELECT ID,TITLE FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='" . UserSyear() . "' AND SCHOOL_ID='" . UserSchool() . "' ORDER BY SORT_ORDER,TITLE" );
 
-DrawHeader(ProgramTitle());
-$categories_RET = DBGet( "SELECT ID,TITLE FROM ATTENDANCE_CODE_CATEGORIES WHERE SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."' ORDER BY SORT_ORDER,TITLE" );
-if ( $_REQUEST['table']=='')
+if ( $_REQUEST['table'] == '' )
+{
 	$_REQUEST['table'] = '0';
-$category_select = "<select name=table onChange='ajaxPostForm(this.form,true);'><option value='0'".($_REQUEST['table']=='0'?' selected':'').">"._('Attendance')."</option>";
-foreach ( (array) $categories_RET as $category)
-	$category_select .= '<option value="'.$category[ID].'"'.(($_REQUEST['table']==$category['ID'])?' selected':'').">".$category['TITLE']."</option>";
+}
+
+$category_select = "<select name=table onChange='ajaxPostForm(this.form,true);'><option value='0'" . ( $_REQUEST['table'] == '0' ? ' selected' : '' ) . ">" . _( 'Attendance' ) . "</option>";
+
+foreach ( (array) $categories_RET as $category )
+{
+	$category_select .= '<option value="' . $category[ID] . '"' . (  ( $_REQUEST['table'] == $category['ID'] ) ? ' selected' : '' ) . ">" . $category['TITLE'] . "</option>";
+}
+
 $category_select .= "</select>";
 
-$QI = DBQuery("SELECT sp.PERIOD_ID,sp.TITLE FROM SCHOOL_PERIODS sp WHERE sp.SCHOOL_ID='".UserSchool()."' AND sp.SYEAR='".UserSyear()."' AND EXISTS (SELECT '' FROM COURSE_PERIODS WHERE SYEAR=sp.SYEAR AND PERIOD_ID=sp.PERIOD_ID AND position('," . $_REQUEST['table'] . ",' IN DOES_ATTENDANCE)>0) ORDER BY sp.SORT_ORDER");
-$periods_RET = DBGet($QI,array(),array('PERIOD_ID'));
+$QI = DBQuery( "SELECT sp.PERIOD_ID,sp.TITLE FROM SCHOOL_PERIODS sp WHERE sp.SCHOOL_ID='" . UserSchool() . "' AND sp.SYEAR='" . UserSyear() . "' AND EXISTS (SELECT '' FROM COURSE_PERIODS WHERE SYEAR=sp.SYEAR AND PERIOD_ID=sp.PERIOD_ID AND position('," . $_REQUEST['table'] . ",' IN DOES_ATTENDANCE)>0) ORDER BY sp.SORT_ORDER" );
+$periods_RET = DBGet( $QI, array(), array( 'PERIOD_ID' ) );
 
-$period_select = "<select name=period onChange='ajaxPostForm(this.form,true);'><option value=''>"._('All')."</option>";
-foreach ( (array) $periods_RET as $id => $period)
-	$period_select .= '<option value="'.$id.'"'.(($_REQUEST['period']==$id)?' selected':'').">".$period[1]['TITLE']."</option>";
+$period_select = "<select name=period onChange='ajaxPostForm(this.form,true);'><option value=''>" . _( 'All' ) . "</option>";
+
+foreach ( (array) $periods_RET as $id => $period )
+{
+	$period_select .= '<option value="' . $id . '"' . (  ( $_REQUEST['period'] == $id ) ? ' selected' : '' ) . ">" . $period[1]['TITLE'] . "</option>";
+}
+
 $period_select .= "</select>";
 
-echo '<form action="Modules.php?modname='.$_REQUEST['modname'].'" method="POST">';
-DrawHeader(PrepareDate($date,'_date',false,array('submit'=>true)).' - '.$period_select);
-DrawHeader('',$category_select);
+echo '<form action="Modules.php?modname=' . $_REQUEST['modname'] . '" method="GET">';
+DrawHeader( PrepareDate( $date, '_date', false, array( 'submit' => true ) ) . ' - ' . $period_select );
+DrawHeader( '', $category_select );
 echo '</form>';
 
 //FJ days numbered
 //FJ multiple school periods for a course period
+
 if ( SchoolInfo( 'NUMBER_DAYS_ROTATION' ) !== null )
 {
 	$sql = "SELECT s.STAFF_ID," . DisplayNameSQL( 's' ) . " AS FULL_NAME,
@@ -85,7 +81,9 @@ if ( SchoolInfo( 'NUMBER_DAYS_ROTATION' ) !== null )
 		AND acc.BLOCK IS NOT NULL
 		AND sp.BLOCK=acc.BLOCK)
 	ORDER BY FULL_NAME";
-} else {
+}
+else
+{
 	$sql = "SELECT s.STAFF_ID," . DisplayNameSQL( 's' ) . " AS FULL_NAME,
 		sp.TITLE,cpsp.PERIOD_ID,cp.TITLE AS CP_TITLE,
 		(SELECT 'Y'
@@ -104,8 +102,8 @@ if ( SchoolInfo( 'NUMBER_DAYS_ROTATION' ) !== null )
 		AND cp.SYEAR='" . UserSyear() . "'
 		AND cp.SCHOOL_ID='" . UserSchool() . "'
 		AND s.PROFILE='teacher'" .
-		( $_REQUEST['period'] ? " AND cpsp.PERIOD_ID='" . $_REQUEST['period'] . "'" : '' ) .
-		" AND acc.CALENDAR_ID=cp.CALENDAR_ID
+	( $_REQUEST['period'] ? " AND cpsp.PERIOD_ID='" . $_REQUEST['period'] . "'" : '' ) .
+	" AND acc.CALENDAR_ID=cp.CALENDAR_ID
 		AND acc.SCHOOL_DATE='" . $date . "'
 		AND acc.SYEAR='" . UserSyear() . "'
 		AND (acc.MINUTES IS NOT NULL AND acc.MINUTES>0)
@@ -126,19 +124,19 @@ if ( ! isset( $_REQUEST['period'] )
 	{
 		$i++;
 
-		$staff_RET[ $i ]['FULL_NAME'] = $periods[1]['FULL_NAME'];
+		$staff_RET[$i]['FULL_NAME'] = $periods[1]['FULL_NAME'];
 
 		foreach ( (array) $periods as $period )
 		{
 			if ( isset( $_REQUEST['_ROSARIO_PDF'] ) )
 			{
-				$staff_RET[ $i ][ $period['PERIOD_ID'] ] = ( $period['COMPLETED'] === 'Y' ?
+				$staff_RET[$i][$period['PERIOD_ID']] = ( $period['COMPLETED'] === 'Y' ?
 					_( 'Yes' ) : _( 'No' ) );
 
 				continue;
 			}
 
-			$staff_RET[ $i ][ $period['PERIOD_ID'] ] = MakeTipMessage(
+			$staff_RET[$i][$period['PERIOD_ID']] = MakeTipMessage(
 				$period['CP_TITLE'],
 				_( 'Course Title' ),
 				button( $period['COMPLETED'] === 'Y' ? 'check' : 'x' )
@@ -150,21 +148,22 @@ if ( ! isset( $_REQUEST['period'] )
 
 	foreach ( (array) $periods_RET as $id => $period )
 	{
-		$columns[ $id ] = $period[1]['TITLE'];
+		$columns[$id] = $period[1]['TITLE'];
 	}
 
 	ListOutput( $staff_RET, $columns, 'Teacher who takes attendance', 'Teachers who take attendance' );
 }
 else
 {
-	$period_title = $periods_RET[ $_REQUEST['period'] ][1]['TITLE'];
+	$period_title = $periods_RET[$_REQUEST['period']][1]['TITLE'];
 
 	// FJ display icon for completed column.
+
 	foreach ( (array) $RET as $staff_id => $periods )
 	{
 		foreach ( (array) $periods as $id => $period )
 		{
-			$RET[ $staff_id ][ $id ]['COMPLETED'] = button( $period['COMPLETED'] === 'Y' ? 'check' : 'x' );
+			$RET[$staff_id][$id]['COMPLETED'] = button( $period['COMPLETED'] === 'Y' ? 'check' : 'x' );
 		}
 	}