+++ modules/Food_Service/BalanceReport.php
@@ -1,25 +1,6 @@
 <?php
 
-// set date
-if ( isset( $_REQUEST['month_date'] )
-	&& isset( $_REQUEST['day_date'] )
-	&& isset( $_REQUEST['year_date'] ) )
-{
-	$date = RequestedDate(
-		$_REQUEST['year_date'],
-		$_REQUEST['month_date'],
-		$_REQUEST['day_date']
-	);
-}
-
-if ( empty( $date ) )
-{
-	$_REQUEST['day_date'] = date('d');
-	$_REQUEST['month_date'] = date('m');
-	$_REQUEST['year_date'] = date('Y');
-
-	$date = $_REQUEST['year_date'] . '-' . $_REQUEST['month_date'] . '-' . $_REQUEST['day_date'];
-}
+$date = RequestedDate( 'date', DBDate(), 'set' );
 
 if ( ! empty( $_REQUEST['type'] ) )
 	$_SESSION['FSA_type'] = $_REQUEST['type'];